pyproject.toml
[project]
name = "cryptohacks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.3.0",
    "pwntools>=4.14.1",
    "pycryptodome>=3.23.0",
    "requests>=2.32.5",
]
